# YJIT: 32 bit comparison on shape id