--- a/mjit_c.rb
+++ b/mjit_c.rb
@@ -158,10 +158,6 @@ module RubyVM::MJIT
     Primitive.cexpr! %q{ INT2NUM(VM_METHOD_TYPE_ISEQ) }
   end
 
-  def C.SHAPE_BITS
-    Primitive.cexpr! %q{ UINT2NUM(SHAPE_BITS) }
-  end
-
   def C.SHAPE_CAPACITY_CHANGE
     Primitive.cexpr! %q{ UINT2NUM(SHAPE_CAPACITY_CHANGE) }
   end
@@ -172,6 +168,10 @@ module RubyVM::MJIT
 
   def C.SHAPE_FROZEN
     Primitive.cexpr! %q{ UINT2NUM(SHAPE_FROZEN) }
+  end
+
+  def C.SHAPE_ID_NUM_BITS
+    Primitive.cexpr! %q{ UINT2NUM(SHAPE_ID_NUM_BITS) }
   end
 
   def C.SHAPE_INITIAL_CAPACITY