@tenderlove 2022-11-16
I ran YJIT Bench with this branch:

```
before: ruby 3.2.0dev (2022-11-16T18:09:15Z master 3eb7a6521c) +YJIT [x86_64-linux]
after: ruby 3.2.0dev (2022-11-15T21:31:44Z 32-bit-compare 962dc0eafb) +YJIT [x86_64-linux]

-------------  -----------  ----------  ----------  ----------  ------------  -------------
bench          before (ms)  stddev (%)  after (ms)  stddev (%)  before/after  after 1st itr
30k_ifelse     222.8        0.0         230.2       0.1         0.97          0.98         
30k_methods    536.6        1.0         559.7       0.0         0.96          0.96         
activerecord   69.3         0.2         69.0        0.2         1.00          1.01         
binarytrees    139.0        2.6         138.9       0.3         1.00          0.99         
cfunc_itself   30.7         0.1         31.4        0.5         0.98          0.97         
chunky_png     443.0        0.4         481.9       0.2         0.92          0.92         
erubi          199.6        0.9         203.3       0.9         0.98          0.98         
erubi_rails    12.4         2.8         12.3        2.8         1.01          0.99         
etanni         339.5        0.4         345.3       0.4         0.98          0.99         
fannkuchredux  1446.7       1.7         1439.4      0.2         1.01          1.00         
fib            41.7         0.2         41.7        0.1         1.00          0.99         
getivar        23.3         0.1         24.3        0.1         0.96          1.01         
hexapdf        1506.0       2.5         1521.7      2.4         0.99          1.01         
keyword_args   39.0         0.3         39.1        0.2         1.00          1.00         
lee            654.4        1.5         651.5       1.3         1.00          1.01         
liquid-render  77.9         1.6         77.1        1.6         1.01          0.99         
mail           106.0        0.2         106.2       0.2         1.00          0.99         
nbody          63.1         0.2         60.3        0.4         1.05          1.05         
optcarrot      2150.5       0.6         2117.3      0.6         1.02          1.01         
psych-load     1355.3       0.1         1351.2      0.1         1.00          1.00         
railsbench     1297.7       1.4         1277.4      1.9         1.02          1.01         
respond_to     18.1         3.5         20.8        0.2         0.87          0.86         
ruby-lsp       51.1         20.5        51.1        20.9        1.00          0.99         
rubykon        4849.6       0.4         4830.2      0.4         1.00          1.00         
setivar        30.3         0.3         31.2        8.1         0.97          1.01         
str_concat     26.5         2.1         26.5        2.3         1.00          0.99         
-------------  -----------  ----------  ----------  ----------  ------------  -------------
Legend:
- before/after: ratio of before/after time. Higher is better for after. Above 1 represents a speedup.
- after 1st itr: ratio of before/after time for the first benchmarking iteration.
```

Looks like `nbody` is faster.  `respond_to` is slower, but I don't understand why.  I looked at the benchmark and it has nothing to do with instance variables, so I don't understand.