@tenderlove 2022-11-17
> I think it's good that the shape id comparison can be made smaller, but the perf drop on `respond_to` is pretty big. It might be worth looking at the disasm. It's possible that the changes to the splitting code may have caused the difference.

I'll take a look at the disasm.  I don't think the splitting impacted it though as I've only messed with splitting on arm64 and the benchmarks were on x86.