@tenderlove 2022-11-18
I've updated this against latest master.  The new machine code is like this:

```
== BLOCK 1/4, ISEQ RANGE [0,3), 41 bytes ======================
  # getinstancevariable
  0x5560c986d023: mov rax, qword ptr [r13 + 0x18]
  # guard object is heap
  0x5560c986d027: test al, 7
  0x5560c986d02a: jne 0x5560c986f02d
  0x5560c986d030: cmp rax, 4
  0x5560c986d034: jbe 0x5560c986f02d
  # guard shape
  0x5560c986d03a: cmp word ptr [rax + 6], 0x19
  0x5560c986d03f: jne 0x5560c986f046
  0x5560c986d045: mov rax, qword ptr [rax + 0x10]
  0x5560c986d049: mov qword ptr [rbx], rax

== BLOCK 2/4, ISEQ RANGE [3,6), 0 bytes =======================
== BLOCK 3/4, ISEQ RANGE [3,6), 23 bytes ======================
  # gen_direct_jmp: fallthrough
  # getinstancevariable
  # regenerate_branch
  # getinstancevariable
  # regenerate_branch
  0x5560c986d04c: mov rax, qword ptr [r13 + 0x18]
  # guard shape
  0x5560c986d050: cmp word ptr [rax + 6], 0x19
  0x5560c986d055: jne 0x5560c986f099
  0x5560c986d05b: mov rax, qword ptr [rax + 0x18]
  0x5560c986d05f: mov qword ptr [rbx + 8], rax
```

Benchmarks are like this:

```
before: ruby 3.2.0dev (2022-11-18T18:03:33Z master ba3b40a9ae) +YJIT [x86_64-linux]
after: ruby 3.2.0dev (2022-11-18T18:28:20Z 32-bit-compare 8190d35c5d) +YJIT [x86_64-linux]

-------------  -----------  ----------  ----------  ----------  ------------  -------------
bench          before (ms)  stddev (%)  after (ms)  stddev (%)  before/after  after 1st itr
30k_ifelse     230.2        0.0         231.1       1.0         1.00          1.00         
30k_methods    560.8        0.0         558.6       0.0         1.00          1.00         
activerecord   69.7         0.2         69.2        0.2         1.01          1.00         
binarytrees    135.9        0.3         139.7       1.6         0.97          0.98         
cfunc_itself   30.5         0.2         30.5        0.1         1.00          1.00         
chunky_png     461.2        0.3         431.8       0.4         1.07          1.06         
erubi          197.7        0.7         197.7       0.9         1.00          1.00         
erubi_rails    12.1         4.0         12.1        3.4         1.00          1.00         
etanni         330.6        0.4         331.8       0.3         1.00          1.00         
fannkuchredux  1394.8       0.2         1400.6      0.6         1.00          1.04         
fib            41.6         0.9         41.5        0.1         1.00          1.01         
getivar        23.3         0.1         24.5        0.1         0.95          1.00         
hexapdf        1457.7       1.5         1460.8      2.1         1.00          1.00         
keyword_args   38.1         0.2         38.2        0.1         1.00          1.00         
lee            624.4        0.8         633.7       1.0         0.99          1.00         
liquid-render  76.1         1.8         76.1        1.6         1.00          1.00         
mail           105.3        0.3         105.5       0.1         1.00          1.00         
nbody          59.4         1.4         58.8        0.4         1.01          1.02         
optcarrot      2155.9       0.6         2018.3      0.6         1.07          1.07         
psych-load     1358.8       0.1         1354.6      0.1         1.00          1.00         
railsbench     1279.3       1.1         1273.8      2.1         1.00          1.00         
respond_to     21.6         0.2         22.0        0.1         0.98          0.99         
ruby-lsp       50.6         19.5        50.3        20.3        1.01          1.00         
rubykon        4756.7       0.4         4832.8      0.4         0.98          0.98         
setivar        30.0         0.3         30.4        1.0         0.99          1.01         
str_concat     26.6         1.6         26.7        2.5         0.99          1.00         
-------------  -----------  ----------  ----------  ----------  ------------  -------------
Legend:
- before/after: ratio of before/after time. Higher is better for after. Above 1 represents a speedup.
- after 1st itr: ratio of before/after time for the first benchmarking iteration.
```